+++ src/Cache/BaseDataCache.php
@@ -80,7 +80,7 @@ final class BaseDataCache implements DataCache
      * @throws InvalidArgumentException
      *   MUST be thrown if the $key string is not a legal value.
      */
-    public function get_data($key, $default = null)
+    public function get_data(string $key, $default = null)
     {
         $data = $this->cache->load();
 
@@ -123,9 +123,9 @@ final class BaseDataCache implements DataCache
      * @throws InvalidArgumentException
      *   MUST be thrown if the $key string is not a legal value.
      */
-    public function set_data($key, array $value, $ttl = null)
+    public function set_data(string $key, array $value, ?int $ttl = null): bool
     {
-        if (! is_int($ttl)) {
+        if ($ttl === null) {
             $ttl = 3600;
         }
 
@@ -150,7 +150,7 @@ final class BaseDataCache implements DataCache
      * @throws InvalidArgumentException
      *   MUST be thrown if the $key string is not a legal value.
      */
-    public function delete_data($key)
+    public function delete_data(string $key): bool
     {
         return $this->cache->unlink();
     }